+++ Demos/src/main/java/com/estimote/examples/demos/AllDemosActivity.java
@@ -42,5 +42,13 @@ public class AllDemosActivity extends Activity {
         startActivity(intent);
       }
     });
+    findViewById(R.id.update_demo_button).setOnClickListener(new View.OnClickListener() {
+      @Override
+      public void onClick(View v) {
+        Intent intent = new Intent(AllDemosActivity.this, ListBeaconsActivity.class);
+        intent.putExtra(ListBeaconsActivity.EXTRAS_TARGET_ACTIVITY, UpdateDemoActivity.class.getName());
+        startActivity(intent);
+      }
+    });
   }
 }